server/requirements.txt
chardet==5.2.0
aiohttp==3.8.1
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
paddlepaddle==2.6.1
paddleocr==2.8.1
Pillow
pytesseract==0.3.10
torch==2.0.1
psd2fabric==0.1
requests~=2.31.0
flask~=3.0.3
werkzeug~=3.0.2
iopaint==1.3.3
oss2
sqlalchemy
pymysql
boto3==1.26.87
click==8.1.3
grpc-interceptor==0.13.0
grpcio==1.51.3
grpcio-health-checking==1.51.3
gunicorn==20.1.0
prometheus-client==0.14.1
protobuf==4.22.0
redis==3.5.3
toml==0.10.1
scikit-learn
websockets
websocket-client
babel
pint
volcengine